sofar_poller/requirements.txt
bitarray==2.9.2
bitstring==4.1.4
paho-mqtt==2.0.0
pip==23.3.1
pyModbusTCP==0.2.1
setuptools==68.2.2
wheel==0.41.2
